extract cors config to another file
